+++ decs.d.ts
@@ -1,7 +1,7 @@
 declare module "evmole";
 
 interface EthereumProvider {
-  request: (args: { method: string, params?: unknown[] }) => Promise<unknown>;
+  request: (args: { method: string; params?: unknown[] }) => Promise<unknown>;
 }
 
 interface Window {