+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 
+- Add decoding of return parameter types.
+
 ## 1.0.0
 
 - Initial `front end` to interact with any contract (no ABI/source coded needed).